+++ commands/__init__.py
@@ -11,4 +11,4 @@ top_level_command_args = {
 
 def get_usage() -> str:
     command_names = list(top_level_command_args.keys())
-    return f"usage: spotiList {{help,{','.join(command_names)}}}"
+    return f"usage: spotiList {{help,configure,{','.join(command_names)}}}"